chore(repo): move test-only workspace deps out of crates
