+++ drivers/rdrive/tests/init_sources.rs
@@ -28,6 +28,21 @@ static STATIC_REGISTER: DriverRegister = DriverRegister {
 };
 
 #[test]
+fn acpi_source_is_unsupported() {
+    let err = rdrive::init(Platform::Acpi(AcpiRoot::identity(0)))
+        .expect_err("invalid acpi should be rejected");
+
+    assert!(matches!(err, rdrive::error::DriverError::Unknown(_)));
+}
+
+#[test]
+fn fdt_phandle_lookup_is_none_without_fdt_source() {
+    init_sources(&[PlatformSource::Static]).expect("static source should init");
+
+    assert!(rdrive::fdt_phandle_to_device_id(1.into()).is_none());
+}
+
+#[test]
 fn unsupported_source_does_not_leave_static_backend_initialized() {
     let err = init_sources(&[
         PlatformSource::Static,
