+++ drivers/test_crates/driver-tests/src/lib.rs
@@ -1,0 +1,4 @@
+//! Non-published manual driver test package.
+//!
+//! The test files in this crate are intentionally not auto-discovered by Cargo.
+//! They require bare-metal runners or board/QEMU setup outside the default CI.
