+++ virtualization/axaddrspace/Cargo.toml
@@ -39,6 +39,5 @@ x86 = "0.52"
 
 [dev-dependencies]
 lazy_static = "1.5"
-ax-kspin.workspace = true
 assert_matches = "1.5.0"
 axin = "0.1.0"
