+++ virtualization/axdevice/Cargo.toml
@@ -30,6 +30,3 @@ arm_vgic = { workspace = true, features = ["vgicv3"] }
 riscv_vplic = { workspace = true }
 [target.'cfg(target_arch = "x86_64")'.dependencies]
 x86_vlapic = { workspace = true }
-
-[dev-dependencies]
-ax-crate-interface = { workspace = true }
