+++ virtualization/test_crates/virtualization-tests/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "virtualization-tests"
+version = "0.1.0"
+edition.workspace = true
+publish = false
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+
+[dev-dependencies]
+ax-crate-interface.workspace = true
+ax-errno.workspace = true
+ax-memory-addr.workspace = true
+axdevice.workspace = true
+axdevice_base.workspace = true
+axvm-types.workspace = true
+x86_vlapic.workspace = true
