+++ virtualization/test_crates/virtualization-tests/src/lib.rs
@@ -1,0 +1,1 @@
+//! Non-published virtualization integration tests.
